requirements.txt
babel==2.16.0
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
pillow==11.0.0
tkcalendar==1.6.1
